network slection in menu moved code from fragment to activity; dropdownlist is now better
