+++ app/src/main/res/layout/login_fragment.xml
@@ -4,41 +4,6 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ToggleButton
-            android:id="@+id/tbMainNet"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:checked="false"
-            android:textOff="@string/connect_testnet"
-            android:textOn="@string/connect_mainnet"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <com.m2049r.xmrwallet.layout.InstantAutoComplete
-            android:id="@+id/etDaemonAddress"
-            style="@style/MoneroEdit"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:clickable="true"
-            android:focusable="true"
-            android:focusableInTouchMode="true"
-            android:gravity="center"
-            android:hint="@string/prompt_daemon"
-            android:imeOptions="actionDone"
-            android:inputType="textWebEmailAddress|textNoSuggestions"
-            android:maxLines="1"
-            android:textIsSelectable="true"
-            app:layout_constraintBaseline_toBaselineOf="@+id/tbMainNet"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toLeftOf="@+id/tbMainNet" />
-
-    </android.support.constraint.ConstraintLayout>
-
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
